packages/edge-auth/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-frontend-edge-auth"
version = "0.1.0"
description = "Pluggable edge authentication for voice endpoints"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1",
]

[project.optional-dependencies]
cloudflare = ["PyJWT>=2.8,<3", "cryptography>=42.0"]
google = ["google-auth>=2.20,<3"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
